facebook_automation/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.4
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
playwright==1.40.0
pyautogui==0.9.54
schedule==1.2.0
logging
json5==0.9.14
configparser==6.0.0
argparse
time
random
os
sys
threading
datetime
